creates composite_shape module
