+++ tests/core/test_assembly.py
@@ -1,7 +1,7 @@
 import unittest
 import volmdlr
 import volmdlr.primitives3d as p3d
-import volmdlr.core
+import volmdlr.composite_shapes
 
 
 class TestAssembly(unittest.TestCase):
@@ -9,7 +9,7 @@ class TestAssembly(unittest.TestCase):
         box1 = p3d.Block(volmdlr.OXYZ)
         components = [box1, box1]
         positions = [volmdlr.OXYZ, volmdlr.Frame3D(volmdlr.Point3D(0, 0, 1), volmdlr.X3D, volmdlr.Y3D, volmdlr.Z3D)]
-        assembly = volmdlr.core.Assembly(components, positions)
+        assembly = volmdlr.composite_shapes.Assembly(components, positions)
         self.assertAlmostEqual(assembly.bounding_box.volume(), 2.0)  # add assertion here
